- Added new folder for presentation - Restructured Gateway into packages
